backend/requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
boto3==1.34.74
botocore==1.34.74
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
flower==2.0.1
humanize==4.9.0
idna==3.6
inflection==0.5.1
jmespath==1.0.1
kombu==5.3.6
packaging==24.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
PyJWT==2.8.0
pyTelegramBotAPI==4.16.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
ruff==0.3.4
s3transfer==0.10.1
setuptools==69.2.0
six==1.16.0
sqlparse==0.4.4
tornado==6.4
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
django-storages==1.14.2